fix: 뒤로 가기 버튼 수정 및 버튼 disable 처리
